+++ package/recalbox-romfs2/scripts/systemholder.py
@@ -25,7 +25,7 @@ class SystemHolder:
             "mandatory": 3,
         }
 
-        def __init__(self, systemtype: str, pad: str, keyboard: str, mouse: str, lightgun: str, releasedate: str):
+        def __init__(self, systemtype: str, pad: str, keyboard: str, mouse: str, lightgun: str, releasedate: str, retroachievements: str):
             if systemtype not in self.__systemTypes.keys(): raise TypeError("Invalid system type! {}".format(systemtype))
             if pad not in self.__deviceRequirement.keys(): raise TypeError("Invalid pad type! {}".format(pad))
             if keyboard not in self.__deviceRequirement.keys(): raise TypeError("Invalid keyboard type! {}".format(keyboard))
@@ -35,8 +35,9 @@ class SystemHolder:
             self.__pad = pad
             self.__keyboard = keyboard
             self.__mouse = mouse
-            self.__lightgun = lightgun
+            self.__lightgun:bool = lightgun
             self.__releasedate = releasedate
+            self.__retroachievements:bool = retroachievements
 
         def serialize(self):
             return {
@@ -44,8 +45,9 @@ class SystemHolder:
                 "pad": self.__pad,
                 "keyboard": self.__keyboard,
                 "mouse": self.__mouse,
-                "lightgun": self.__lightgun,
-                "releasedate": self.__releasedate
+                "lightgun": '0' if self.__lightgun == 'no' else '1',
+                "releasedate": self.__releasedate,
+                "retroachievements": '1' if self.__retroachievements else '0'
             }
 
         @property
@@ -73,7 +75,13 @@ class SystemHolder:
         def MouseEnum(self): return self.__deviceRequirement[self.__mouse]
 
         @property
+        def LightGun(self) -> bool: return self.__lightgun
+
+        @property
         def ReleaseDate(self): return self.__releasedate
+
+        @property
+        def RetroAchievements(self) -> bool: return self.__retroachievements
 
     class Core:
 
@@ -161,7 +169,7 @@ class SystemHolder:
         }
         self.__port: bool = False
         self.__readOnly: bool = False
-        self.__properties = SystemHolder.SystemProperties("virtual", "no", "no", "no", "no", "")
+        self.__properties = SystemHolder.SystemProperties("virtual", "no", "no", "no", "no", "", "")
         self.__coreLists: Dict[str, List[SystemHolder.Core]] = dict()
         self.__coreCount: int = 0
         self.__deserialize()
@@ -297,5 +305,6 @@ class SystemHolder:
             mouse=self.__get(desc, "properties", "device.mouse", "no", True),
             lightgun=self.__get(desc, "properties", "device.lightgun", "no", True),
             releasedate=self.__get(desc, "properties", "release.date", "", True),
+            retroachievements=(self.__get(desc, "properties", "retroachievements", "", True) == '1'),
         )